Moved custom commands to a separate file
